Moved models from test_multi.py to test_models.py #45
* Previous commit did not trigger CI build since I omitted the PR reference this commit should do the trick
